RUB-280 rust: split DA set validation helpers
